에러 수정 및 과제 이미지

+++ Homework3/src/main/java/kr/ac/hansung/cse/controller/ProductController.java
@@ -56,7 +56,7 @@ public class ProductController {
 	}
 	
 	@GetMapping("/products/{id}")
-	public ResponseEntity<Product> getCustomerById(@PathVariable("id") long id) {
+	public ResponseEntity<Product> getCustomerById(@PathVariable("id") int id) {
 		Optional<Product> productData = repository.findById(id);
 
 		if (productData.isPresent()) {
@@ -82,7 +82,7 @@ public class ProductController {
 	
 	
 	@PutMapping("/products/{id}")
-	public ResponseEntity<Product> updateProduct(@PathVariable("id") long id, @RequestBody Product product) {
+	public ResponseEntity<Product> updateProduct(@PathVariable("id") int id, @RequestBody Product product) {
 		Optional<Product> productData = repository.findById(id);
 
 		if (productData.isPresent()) {
@@ -102,7 +102,7 @@ public class ProductController {
 	
 	
 	@DeleteMapping("/products/{id}")
-	public ResponseEntity<HttpStatus> deleteCustomer(@PathVariable("id") long id) {
+	public ResponseEntity<HttpStatus> deleteCustomer(@PathVariable("id") int id) {
 		try {
 			repository.deleteById(id);
 			return new ResponseEntity<>(HttpStatus.NO_CONTENT);
